refactor(folder structure):create components,page structure and move file in folder
